+++ src/app/components/editors/ck-editor/example/ck-editor-directive-example.component.ts
@@ -1,0 +1,13 @@
+import { Component } from '@angular/core';
+import { ckEditorMinimumConfig } from '../presets/ck-editor-minimum-preset';
+import { ckEditorFullConfig } from '../presets/ck-editor-full-preset';
+
+@Component({
+    selector: 'tc-ck-editor-directive-example',
+    template: require('./ck-editor-directive-example.component.html')
+})
+export class CkEditorDirectiveExample
+{
+    protected config:{} = ckEditorFullConfig;
+    protected tiny:{} = ckEditorMinimumConfig;
+}