+++ src/app/components/interactables/draggable.directive.ts
@@ -23,7 +23,8 @@ import interact = require('interactjs');
 })
 export class TerraDraggableDirective implements OnInit, OnChanges
 {
-    @Input()
+    /* tslint:disable-next-line:no-input-rename */
+    @Input('terraDraggable')
     public options?:DraggableOptions = null;
 
     @Input()
@@ -51,16 +52,6 @@ export class TerraDraggableDirective implements OnInit, OnChanges
     public end:EventEmitter<InteractEvent> = new EventEmitter<InteractEvent>();
 
     /* tslint:disable:no-output-on-prefix no-input-rename no-output-rename */
-    /**
-     * @deprecated since 3.x.x. Use options instead.
-     */
-    @Input('terraDraggable')
-    public set draggableOptions(value:DraggableOptions)
-    {
-        console.warn('`terraDraggable` is deprecated. Please use `options` instead.');
-        this.options = value;
-    }
-
     /**
      * @deprecated since 3.x.x. Use disabled instead.
      */
@@ -227,16 +218,19 @@ export class TerraDraggableDirective implements OnInit, OnChanges
             onstart:      (event:InteractEvent):void =>
                           {
                               this.onStart.emit(event);
+                              this.start.emit(event);
                               event.target.IA_DRAG_DATA = this.dragData;
                           },
             onmove:       (event:InteractEvent):void =>
                           {
                               this.onMove.emit(event);
+                              this.move.emit(event);
                               event.target.IA_DRAG_DATA = this.dragData;
                           },
             onend:        (event:InteractEvent):void =>
                           {
                               this.onEnd.emit(event);
+                              this.end.emit(event);
                               event.target.IA_DRAG_DATA = null;
                           },
         };