+++ CHANGELOG.md
@@ -1,3 +1,37 @@
+<a name="4.0.0-beta.5"></a>
+# 4.0.0-beta.5 (28.06.2019)
+
+### Breaking Changes
+* **TerraComponentsModule** removed `forRoot()` and `forChild()` methods since all services use tree-shakable providers now.
+
+### Bug Fixes
+* **Tooltip** fixed tooltip position when `container="'body'"`.
+* **terra-portlet** fixed orientation of the collapsed icon.
+
+<a name="4.0.0-beta.4"></a>
+# 4.0.0-beta.4 (24.06.2019)
+
+### Breaking Changes
+* **bootstrap** updated from 4.0.0-alpha.4 to 4.3.1
+
+### Features
+* **tcCkEditor** New directive to provide a full config for the ckeditor4.
+* **ck-editor-presets** Config presets for the ckeditor4.
+* **refactoring** restructured style files
+* **terra-cards** restyling
+* **fonts** plentyicon-font moved to aws S3
+
+<a name="4.0.0-beta.3"></a>
+# 4.0.0-beta.3 (12.06.2019)
+
+### Bug Fixes
+* **button-group** fixed spacing between buttons in a button group
+* **angular-l10n** added missing `OnInit` and `OnDestroy` LifeCycle-Hooks to Components using angular-l10n's `Language` decorator
+
+<a name="4.0.0-beta.2"></a>
+# 4.0.0-beta.2 (11.06.2019)
+* **bootstrap** revert bootstrap update
+
 <a name="4.0.0-beta.1"></a>
 # 4.0.0-beta.1 (07.06.2019)
 
@@ -18,9 +52,12 @@
 
 ### Features
 * **tslint** add new linting rules
+* **terra-category-picker** allow to accept 'Null' values
+* **terra-cacheckbox** allow to accept 'Null' values
 
 ### Bug Fixes
 * **http-interceptors** fix logging errors to console. 
+* **draggable** allow `terraDraggable` input to be aliased.
 
 <a name="3.14.0"></a>
 # 3.14.0 (29.05.2019)